requirements.txt
Flask
easyocr
Pillow
